use existing query-string-depth config to control expr recursion limit
