+++ src/expr/expr.cc
@@ -14,6 +14,7 @@
 
 #include "absl/strings/str_cat.h"
 #include "src/utils/scanner.h"
+#include "src/valkey_search_options.h"
 #include "vmsdk/src/status/status_macros.h"
 #include "vmsdk/src/valkey_module_api/valkey_module.h"
 
@@ -280,9 +281,14 @@ bool IsIdentifierChar(int c) {
   return c != EOF && (std::isalnum(c) || c == '_');
 }
 
+struct DepthGuard {
+  int& depth;
+  DepthGuard(int& d) : depth(d) { ++depth; }
+  ~DepthGuard() { --depth; }
+};
+
 struct Compiler {
   int depth_ = 0;
-  static constexpr int kMaxDepth = 1000;
   utils::Scanner s_;
   Compiler(absl::string_view sv) : s_(sv) {}
 
@@ -351,7 +357,8 @@ struct Compiler {
   }
 
   absl::StatusOr<ExprPtr> Primary(CompileContext& ctx) {
-    if (++depth_ > kMaxDepth) {
+    DepthGuard guard(depth_);
+    if (depth_ > options::GetQueryStringDepth().GetValue()) {
       return absl::InvalidArgumentError("Expression too complex");
     }
     s_.SkipWhiteSpace();
